requirements.txt
Django==1.6.5
django-admin-sortable2==0.2.7
django-localflavor==1.0
